cleaning up + teams 0.5
