+++ readme.md
@@ -1,7 +1,7 @@
 СИСАДМИН_ОЧКА
 ---------------------
 ## Эндпоинты
-#### СисадминОчка
+### СисадминОчка
  - **POST /users** - регистрация нового пользователя, требует *name*, *email*, *position* и *password*(7-20символов) в json в body, выдает ***auth-token*** в респонс, пользователь остается авторизованным 360000000 секунд(дев).
  - **POST /auth** - авторизация, требует *email* и *password* ранее зарегестрированного пользователя в json в body, выдает ***auth-token*** в респонс, пользователь остается авторизованным 360000000 секунд(дев).
   - **GET /tickets/all** - отправляет в респонс массив со всеми проектами, отсортированными по дате.
@@ -12,11 +12,12 @@
   - **GET /tickets/all/emergency** - показывает все тикеты, отсортированные по emergency.
   - **PUT /tickets/id** - меняет *status* тикета на *false*, ~закрывая его.
 
-#### ПроектыОчка
-  - **POST /** :joy: - добавить новый проект, указав *name*, *dateStart*, *city*, *type*, *stage* в body
+### ПроектыОчка
+  - **POST /projects/add** :joy: - добавить новый проект, указав *name*, *dateStart*, *city*, *type*, *stage* в body
   - **GET /projects** :star_struck: - получить список всех проектов
   - **GET /projects/:auth** :sunglasses: - найти проект по *auth*, где *auth* это *crypt* или *name* проекта, указанные в хедере, в респонсе выдает инфу по конкретному проекту если указан crypt, или все проекты с указанным name, если искали через него
   - **DELETE /projects/:crypt** :grimacing: - удалить проект, указав его *crypt* в хедере
   - **PUT /projects/:crypt** :stuck_out_tongue_winking_eye: - изменить *name* и *date* проекта, указав его crypt в хедере и новые значения в body. В респонсе измененные данные проекта
-  - **GET /city/:city** :cowboy_hat_face: - респонс со всеми проектами в городе, указанном в *city*
+  - **GET projects/city/:city** :cowboy_hat_face: - респонс со всеми проектами в городе, указанном в *city*
   - **GET projects/user/:id** - найти все проекты пользователя *id*
+  - **PUT projects/updteam/:crypt** - находит проект по его *crypt* и добавляет в его команду пользователя, чей айди указан в *userid* в body
